% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_sunburst.R
\name{count_to_sunburst}
\alias{count_to_sunburst}
\title{Create a sunburst plot from count data}
\usage{
count_to_sunburst(count_df)
}
\arguments{
\item{count_df}{}
}
\value{

}
\description{
Create a sunburst plot from count data
}
